rename HttpServer.cpp to Swiftly.cpp
Previously HttpServer was defined in Swiftly.h and implemented in HttpServer.cpp

This works with qmake, but when building with cmake, there is some linking errors when a project wants to link libSwiftly.a

I figured out that the reason is that when named differently, cmake won't generate the right moc files.

It seems to be a bug of cmake, but for now, I have to solve this by using matching file names.